added complexity-penalty, added smoothing to scatterplots
