+++ trained_models/C_0_cps.pth
