+++ trained_models/C_0_exp.pth
