+++ trained_models/C_1_cps.pth
