+++ trained_models/G_0_cps.pth
