+++ trained_models/G_0_exp.pth
