+++ trained_models/G_1_cps.pth
