+++ trained_models/G_1_exp.pth
